hawkbit: update docs / jar to hawkBit Update Server 0.2.1
Latest release of the hawkBit Update Server is 0.2.1.  Currently,
we are pulling 0.2.0M4 due to a change in 0.2.0M5 which removed
anonymous access to the DDI API.  Also, the Management API was
updated in a way which broke our hawkbit.py management script.

Now that the hawkbit.py script has been updated, let's include
a custom build of 0.2.1 here which re-enables anonymous DDI API
access and update the README with instructions for re-creating it.

Signed-off-by: Michael Scott <mike@foundries.io>